Bugfix: Fixed blankshields->shields bug
There was a bug where characters with a blankshield would sometimes be
generated with a shield. This was caused by the shield name not being
set manually on a blankshield so the DOM wasn't being set. This let
characters inherit shields from previously generated characters. Fixed
by manually passing empty strings when there is a blank shield.