Fixed bugs and improved mobile scaling
Fixed a bug with the user deselecting all of their races or classes in
the options modal. Now when this happens a pretty modal informs the user
of their error and how to correct it.

Improved mobile scaling of the options modal and the new pretty modal
with media queries so that they should be more usable on mobile devices.